requirement.txt
scrapy-rotating-proxies
